// src/lib/enums/transactions/transaction-log-action.enum.ts
export enum TransactionLogAction {
  TRANSACTION_CREATED = 'TRANSACTION_CREATED',
  SUBMIT_TO_UPSTREAM = 'SUBMIT_TO_UPSTREAM',
  UPSTREAM_RESPONSE = 'UPSTREAM_RESPONSE',
  TRANSACTION_SUCCESS = 'TRANSACTION_SUCCESS',
  TRANSACTION_FAILED = 'TRANSACTION_FAILED',
  SET_TO_FAILED_BY_ADMIN = 'SET_TO_FAILED_BY_ADMIN',
  FAILED_ADDED_TO_RETRY_QUEUE = 'FAILED_ADDED_TO_RETRY_QUEUE',
  SEND_NOTIFICATION_TO_MERCHANT = 'SEND_NOTIFICATION_TO_MERCHANT',
  MANUALLY_TRIGGER_NOTIFICATION_TO_MERCHANT = 'MANUALLY_TRIGGER_NOTIFICATION_TO_MERCHANT',
  REVENUE_DISTRIBUTED = 'REVENUE_DISTRIBUTED',
}